lang-graph/requirements.txt
langgraph>=0.4.0
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-ollama>=0.3.0
langchain-core>=0.3.0
python-dotenv>=1.1.0
httpx>=0.28.0
pydantic>=2.10.0
